requirements.txt
python-telegram-bot==21.10
elevenlabs==1.50.5
python-dotenv==1.1.0
